# artifacts/api-server/src/routes/analyst/models.ts
/**
 * Model training, validation, promotion and the kill switch.
 *
 * Split out of routes/analyst.ts by remediation task 5.2. A PURE MOVE: every
 * handler below is byte-identical to the one that was in that file, mounted at
 * the same absolute path. No behaviour changed in the same commit.
 *
 * Routes in this module, in registration order:
 *   POST   /analyst/models/train
 *   GET    /analyst/models
 *   POST   /analyst/models/validate
 *   GET    /analyst/models/validation
 *   POST   /analyst/models/:versionId/promote
 *   POST   /analyst/models/:versionId/demote
 */
import { Router, type IRouter, type RequestHandler } from "express";
import { createHmac, timingSafeEqual } from "node:crypto";
import {
  GetAnalystBullpenRoomResponse,
  GetAnalystDataHealthResponse,
  GetAnalystGameLabResponse,
  GetAnalystMarketResearchResponse,
  GetAnalystPitcherLabResponse,
  GetAnalystPlayerLabResponse,
  GetAnalystProjectionsResponse,
  GetAnalystSettingsResponse,
  GetAnalystTodayResponse,
  CorrectFeatureStoreSnapshotBody,
  RefreshAnalystResearchResponse,
  GetAnalystBatterPitcherResponse,
  RefreshAnalystBatterPitcherResponse,
  GetAnalystRoundRobinComparisonResponse,
  RefreshBullpenResponse,
  RefreshMarketResearchTBResponse,
  RefreshMarketResearchXBHResponse,
  RefreshMarketResearchWALKResponse,
  RefreshMarketResearchHRResponse,
  WriteFeatureStoreOutcomeBody,
  TrainAnalystModelResponse,
  DemoteAnalystModelBody,
  DemoteAnalystModelResponse,
  GetAnalystModelsResponse,
  PromoteAnalystModelResponse,
  ValidateAnalystModelResponse,
  GetAnalystModelValidationResponse,
  GetAnalystDailyMarketBoardResponse,
  GetAnalystDailyBoardGameSummaryResponse,
  RefreshAnalystDailyMarketBoardResponse,
  GetAnalystBettorSourcesResponse,
  CreateAnalystBettorSourceBody,
  CreateAnalystBettorSourceResponse,
  UpdateAnalystBettorSourceBody,
  UpdateAnalystBettorSourceParams,
  UpdateAnalystBettorSourceResponse,
  DeleteAnalystBettorSourceParams,
  DeleteAnalystBettorSourceResponse,
  IngestAnalystBettorPickBody,
  IngestAnalystBettorPickResponse,
  GetAnalystBettorPicksResponse,
  GetAnalystBettorEvaluationQueryParams,
  GetAnalystBettorEvaluationResponse,
  GetAnalystAiToolRegistryResponse,
  CallAnalystAiToolBody,
  CallAnalystAiToolResponse,
  ChatWithAnalystAiBody,
  ChatWithAnalystAiResponse,
  GetAnalystAiDraftsQueryParams,
  GetAnalystAiDraftsResponse,
  CreateAnalystAiDraftBody,
  CreateAnalystAiDraftResponse,
  ApproveAnalystAiDraftParams,
  ApproveAnalystAiDraftBody,
  ApproveAnalystAiDraftResponse,
  RejectAnalystAiDraftParams,
  RejectAnalystAiDraftBody,
  RejectAnalystAiDraftResponse,
  GetAnalystAiSourcingRegisterQueryParams,
  GetAnalystAiSourcingRegisterResponse,
  DecideAnalystAiSourcingClaimParams,
  DecideAnalystAiSourcingClaimBody,
  DecideAnalystAiSourcingClaimResponse,
  GetAnalystAiResearchNotesQueryParams,
  GetAnalystAiResearchNotesResponse,
  getMarketResearchSelectionEligibility,
} from "@workspace/api-zod";
import { pool } from "@workspace/db";
import { ingestFantasyPros, ingestMlbOfficial } from "../../services/data-foundation";
import { getPitcherLab, getPlayerLab, ingestResearch, ingestStatcastHandednessFallback, researchHealth } from "../../services/research-foundation";
import { getBatterPitcherEvidence, refreshBatterPitcherSlate, type BvpMarket } from "../../services/batter-pitcher-research";
import { compareRoundRobinGame, type RoundRobinBoardId, type RoundRobinCandidate } from "../../services/round-robin-comparison";
import { LINEUP_SOURCE_PRECEDENCE, lineupSourceFilter } from "../../services/lineup-sources";
import { RR_DB_TO_MARKET, RR_MARKET_TO_DB } from "../../services/market-codes";
import { getBullpenRoom, refreshBullpen } from "../../services/bullpen-foundation";
import { runTBEngine } from "../../services/tb-engine";
import { runXBHEngine } from "../../services/xbh-engine";
import { runWALKEngine } from "../../services/walk-engine";
import { runHREngine } from "../../services/hr-engine";
import {
  backfillHistoricalSnapshots,
  captureSlateSnapshots,
  correctSnapshot,
  FeatureStoreValidationError,
  queryFeatureStore,
  writeHistoricalOutcome,
} from "../../services/feature-store";
import {
  createMarketPostmortem,
  queryMarketPostmortems,
  querySettlements,
  settleOfficialDate,
  settleOfficialGame,
  SettlementValidationError,
} from "../../services/settlement";
import {
  MODEL_MARKETS,
  ModelTrainingValidationError,
  queryModelVersions,
  trainMarketModel,
  type ModelMarket,
} from "../../services/model-training";
import {
  queryWalkForwardRuns,
  validateModelVersion,
  WalkForwardValidationError,
} from "../../services/walk-forward-validation";
import {
  demoteModelVersion,
  ModelPromotionError,
  promoteModelVersion,
} from "../../services/model-promotion";
import {
  DailyMarketBoardValidationError,
  type BoardMarket,
  populateDailyMarketBoard,
  queryDailyBoardGameSummary,
  queryDailyMarketBoard,
} from "../../services/daily-market-board";
import {
  automateSettlementDate,
  detectLateScratches,
  interruptOrchestrationRun,
  launchOrchestrationRun,
  queryOrchestrationRuns,
} from "../../services/orchestration";
import { buildSlateExport, buildWorkbookExport } from "../../services/exports";
import { queryAuditEvents } from "../../services/audit";
import { CACHE_POLICY, invalidateCache, readThroughCache } from "../../services/cache";
import {
  BETTOR_MARKETS,
  BettorIntelligenceConflictError,
  BettorIntelligenceValidationError,
  createBettorSource,
  deleteBettorSource,
  ingestBettorPick,
  queryBettorEvaluation,
  queryBettorPicks,
  queryBettorSources,
  updateBettorSource,
  type BettorMarket,
} from "../../services/bettor-intelligence";
import { executeAiToolCall, queryAiToolRegistry, rejectAiToolCall } from "../../services/ai-tool-gateway";
import {
  AiWorkflowValidationError,
  createAiResearchDraft,
  decideAiSourcingClaim,
  queryAiResearchDrafts,
  queryAiSourcingRegister,
  queryResearchNotes,
  reviewAiResearchDraft,
  runAiAnalystChat,
} from "../../services/ai-workflows";
import {
  requestedModelMarket,
} from "./shared";

const router: IRouter = Router();

router.post("/analyst/models/train", async (req, res, next) => {
  try {
    const result = await trainMarketModel(requestedModelMarket(req.query.market));
    res.status(201).json(TrainAnalystModelResponse.parse(result));
  } catch (error) {
    if (error instanceof ModelTrainingValidationError) {
      res.status(400).json({ error: error.message });
      return;
    }
    next(error);
  }
});

router.get("/analyst/models", async (req, res, next) => {
  try {
    const market = req.query.market == null ? null : requestedModelMarket(req.query.market);
    const versions = await queryModelVersions(market);
    res.json(GetAnalystModelsResponse.parse({ versions, total: versions.length }));
  } catch (error) {
    if (error instanceof ModelTrainingValidationError) {
      res.status(400).json({ error: error.message });
      return;
    }
    next(error);
  }
});

router.post("/analyst/models/validate", async (req, res, next) => {
  try {
    const modelVersionId = typeof req.query.modelVersionId === "string"
      ? req.query.modelVersionId.trim()
      : "";
    if (!modelVersionId) throw new WalkForwardValidationError("modelVersionId is required");
    const result = await validateModelVersion(modelVersionId);
    res.status(201).json(ValidateAnalystModelResponse.parse(result));
  } catch (error) {
    if (error instanceof WalkForwardValidationError) {
      res.status(400).json({ error: error.message });
      return;
    }
    next(error);
  }
});

router.get("/analyst/models/validation", async (req, res, next) => {
  try {
    const market = req.query.market == null ? null : requestedModelMarket(req.query.market);
    const runs = await queryWalkForwardRuns(market);
    res.json(GetAnalystModelValidationResponse.parse({ runs, total: runs.length }));
  } catch (error) {
    if (error instanceof WalkForwardValidationError) {
      res.status(400).json({ error: error.message });
      return;
    }
    next(error);
  }
});

/**
 * The only production path to ACTIVE.
 *
 * Deliberately operator-initiated and deliberately not reachable from the
 * orchestration pipeline: promotion stays a human decision until at least two
 * weeks of validation history exist.
 */
router.post("/analyst/models/:versionId/promote", async (req, res, next) => {
  try {
    const result = await promoteModelVersion(req.params.versionId, {
      actor: typeof req.headers["x-operator"] === "string" ? req.headers["x-operator"] : "OPERATOR",
      requestId: typeof req.headers["x-request-id"] === "string" ? req.headers["x-request-id"] : null,
    });
    res.json(PromoteAnalystModelResponse.parse(result));
  } catch (error) {
    if (error instanceof ModelPromotionError) {
      res.status(error.reason === "VERSION_NOT_FOUND" ? 404 : 400)
        .json({ error: error.message, reason: error.reason });
      return;
    }
    next(error);
  }
});

/** The kill switch. One call returns a market to research-only. */
router.post("/analyst/models/:versionId/demote", async (req, res, next) => {
  try {
    const body = DemoteAnalystModelBody.parse(req.body ?? {});
    const result = await demoteModelVersion(req.params.versionId, {
      actor: typeof req.headers["x-operator"] === "string" ? req.headers["x-operator"] : "OPERATOR",
      requestId: typeof req.headers["x-request-id"] === "string" ? req.headers["x-request-id"] : null,
      reason: body.reason ?? null,
    });
    res.json(DemoteAnalystModelResponse.parse(result));
  } catch (error) {
    if (error instanceof ModelPromotionError) {
      res.status(error.reason === "VERSION_NOT_FOUND" ? 404 : 400)
        .json({ error: error.message, reason: error.reason });
      return;
    }
    next(error);
  }
});

export default router;