# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analyst/settlements/{gamePk}:
    post:
      operationId: settleOfficialGame
      tags: [analyst]
      summary: Settle one completed game from MLB Stats API
      parameters:
        - name: gamePk
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "201":
          description: Official game settlement result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfficialGameSettlementResult"
        "400":
          description: Invalid game or non-official settlement request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
                required: [error]
...
  /analyst/models/train:
    post:
      operationId: trainAnalystModel
      tags: [analyst]
      summary: Train one independent market model
      description: |
        Trains a deterministic, market-specific candidate from frozen pregame feature
        snapshots joined only to current official MLB settled outcomes. The artifact is
        stored in private App Storage and pinned by content hash. Training does not
        publish predictions to the market board or activate the model.
      parameters:
        - name: market
          in: query
          required: true
          schema: { type: string, enum: [TB, XBH, WALK, HR] }
      responses:
        "201":
          description: Candidate model training completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelTrainingResult"
        "400":
          description: Invalid market or no official settled training rows
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
                required: [error]
  /analyst/models:
    get:
      operationId: getAnalystModels
      tags: [analyst]
      summary: List versioned market model candidates
      parameters:
        - name: market
          in: query
          required: false
          schema: { type: string, enum: [TB, XBH, WALK, HR] }
      responses:
        "200":
          description: Model versions and lifecycle state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelVersionList"
        "400":
          description: Invalid market
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
                required: [error]
  /analyst/models/{versionId}/promote:
    post:
      operationId: promoteAnalystModel
      tags: [analyst]
      summary: Promote one validated candidate model to ACTIVE
      description: |
        Operator-initiated promotion. Refuses unless the version is a CANDIDATE with a
        walk-forward acceptance whose run passed, met the expected calibration error
        threshold, met the sharpness guard, beat its benchmark by the required margin,
        and carries accepted calibration parameters. Retires whatever was ACTIVE for the
        same market in the same transaction, so exactly one model per market is live.
        Promotion is never performed by the orchestration pipeline.
      parameters:
        - name: versionId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: The version is now ACTIVE
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelPromotionResult"
        "400":
          description: The version does not meet the promotion conditions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelPromotionRefusal"
        "404":
          description: The version was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelPromotionRefusal"
  /analyst/models/{versionId}/demote:
    post:
      operationId: demoteAnalystModel
      tags: [analyst]
      summary: Return one market to research-only
      description: |
        The kill switch. Retires the ACTIVE model for its market, so the next board
        refresh emits RESEARCH_ONLY rows for that market.
      parameters:
        - name: versionId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelDemotionInput"
      responses:
        "200":
          description: The version is retired and the market is research-only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelDemotionResult"
        "400":
          description: The version is not ACTIVE
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelPromotionRefusal"
        "404":
          description: The version was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelPromotionRefusal"
  /analyst/models/validate:
    post:
      operationId: validateAnalystModel
      tags: [analyst]
      summary: Run chronological walk-forward validation for a model version
      description: |
        Evaluates a model using only frozen snapshots and official settled outcomes
        dated before each fold's test date. The run compares out-of-sample Brier skill
        with a market-specific historical base-rate benchmark and performs fold-local
        Platt calibration. A PASS run is required before ACTIVE.
      parameters:
        - name: modelVersionId
          in: query
          required: true
          schema: { type: string }
      responses:
        "201":
          description: Walk-forward validation completed or recorded as incomplete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalkForwardValidationResult"
        "400":
          description: Invalid or unavailable model version
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
                required: [error]
  /analyst/models/validation:
    get:
      operationId: getAnalystModelValidation
      tags: [analyst]
      summary: List walk-forward validation history
      parameters:
        - name: market
          in: query
          required: false
          schema: { type: string, enum: [TB, XBH, WALK, HR] }
      responses:
        "200":
          description: Walk-forward validation runs by market
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalkForwardValidationList"
        "400":
          description: Invalid market
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
                required: [error]
...
  /analyst/refresh/bullpen:
    post:
      operationId: refreshBullpen
      tags: [analyst]
      summary: Ingest reliever appearances for the last 3 days and recompute availability and leverage maps
      parameters:
        - name: date
          in: query
          required: false
          schema: { type: string, format: date }
      responses:
        "201":
          description: Bullpen ingest completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BullpenIngestResult"
components:
  schemas:
...
    BullpenIngestResult:
      type: object
      properties:
        source: { type: string }
        slateDate: { type: string }
        gamesProcessed: { type: integer }
        appearancesNormalized: { type: integer }
        appearancesRejected: { type: integer }
        teamsComputed: { type: integer }
        error: { type: ["string", "null"] }
      required: [source, slateDate, gamesProcessed, appearancesNormalized, appearancesRejected, teamsComputed, error]
...
    UnsettleableBoardCandidate:
      type: object
      properties:
        playerId: { type: integer }
        market: { type: string }
        reason: { type: string }
        gamePk: { type: integer }
      required: [playerId, market, reason]
    SettlementReconciliation:
      type: object
      description: |
        Every board candidate either settled or appears by name in unsettleable
        with the reason it did not. Settlement previously iterated only frozen
        feature snapshots, so a candidate that reached the board without a
        snapshot was never settled and vanished from the record silently.
      properties:
        boardCandidates: { type: integer, minimum: 0 }
        settled: { type: integer, minimum: 0 }
        settledWithoutSnapshot:
          type: integer
          minimum: 0
          description: |
            Settled from the board with no frozen snapshot. A complete
            operational settle record, deliberately excluded from model
            training because it has no pregame feature vector.
        unsettleable:
          type: array
          items: { $ref: "#/components/schemas/UnsettleableBoardCandidate" }
        totalBasesDiscrepancies:
          type: array
          description: |
            Rows where the total bases computed from the components and the
            total bases the feed reported disagree. Settled DISPUTED rather
            than silently taking one of the two.
          items: { type: string }
        walkDefinition:
          type: string
          description: The walk definition applied to this settlement, e.g. BB+IBB (assumed).
        walkDefinitionAssumed: { type: boolean }
        walkDefinitionStatement: { type: string }
      required: [boardCandidates, settled, settledWithoutSnapshot, unsettleable, totalBasesDiscrepancies, walkDefinition, walkDefinitionAssumed, walkDefinitionStatement]
    OfficialGameSettlementResult:
      type: object
      properties:
        gamePk: { type: integer, minimum: 1 }
        slateDate: { type: string, format: date }
        ingestRunId: { type: string }
        state: { type: string, enum: [PENDING, SETTLED, POSTPONED, NO_ACTION, DISPUTED] }
        lines: { type: integer, minimum: 0 }
        outcomesWritten: { type: integer, minimum: 0 }
        corrections: { type: integer, minimum: 0 }
        reconciliation: { $ref: "#/components/schemas/SettlementReconciliation" }
      required: [gamePk, slateDate, ingestRunId, state, lines, outcomesWritten, corrections]
...
    ModelTrainingResult:
      type: object
      additionalProperties: false
      properties:
        trainingRunId: { type: string }
        versionId: { type: string }
        market: { type: string, enum: [TB, XBH, WALK, HR] }
        status: { type: string, enum: [CANDIDATE] }
        trainingSampleCount: { type: integer, minimum: 1 }
        trainingSeasons:
          type: array
          items: { type: string }
        featureSetHash: { type: string }
        algorithm: { type: string }
        artifactKey: { type: string }
        artifactGeneration: { type: string }
        artifactContentHash: { type: string }
      required: [trainingRunId, versionId, market, status, trainingSampleCount, trainingSeasons, featureSetHash, algorithm, artifactKey, artifactGeneration, artifactContentHash]
    ModelVersion:
      type: object
      additionalProperties: false
      properties:
        versionId: { type: string }
        market: { type: string, enum: [TB, XBH, WALK, HR] }
        trainedAt: { type: string, format: date-time }
        trainingSeasons:
          type: array
          items: { type: string }
        featureSetHash: { type: string }
        algorithm: { type: string }
        hyperparameters: { type: object }
        trainingSampleCount: { type: integer, minimum: 1 }
        status: { type: string, enum: [DRAFT, CANDIDATE, ACTIVE, RETIRED, FAILED] }
        artifactKey: { type: string }
        artifactGeneration: { type: string }
        artifactContentHash: { type: string }
        walkForwardAcceptanceId: { type: ["string", "null"] }
        calibrationMethod: { type: ["string", "null"] }
        calibrationSlope: { type: ["number", "null"] }
        calibrationIntercept: { type: ["number", "null"] }
      required: [versionId, market, trainedAt, trainingSeasons, featureSetHash, algorithm, hyperparameters, trainingSampleCount, status, artifactKey, artifactGeneration, artifactContentHash, walkForwardAcceptanceId, calibrationMethod, calibrationSlope, calibrationIntercept]
    ModelVersionList:
      type: object
      additionalProperties: false
      properties:
        versions:
          type: array
          items: { $ref: "#/components/schemas/ModelVersion" }
        total: { type: integer, minimum: 0 }
      required: [versions, total]
    ModelDemotionInput:
      type: object
      additionalProperties: false
      properties:
        reason:
          type: string
          description: Why the market is being returned to research-only. Recorded on the audit event.
    ModelPromotionRefusal:
      type: object
      additionalProperties: false
      properties:
        error: { type: string }
        reason:
          type: string
          description: The single named condition that was not met.
      required: [error, reason]
    ModelPromotionResult:
      type: object
      additionalProperties: false
      properties:
        versionId: { type: string }
        market: { type: string, enum: [TB, XBH, WALK, HR] }
        status: { type: string, enum: [ACTIVE] }
        displacedVersionId: { type: ["string", "null"] }
        displacedStatus: { type: ["string", "null"] }
        walkForwardRunId: { type: string }
        foldCount: { type: integer, minimum: 0 }
        expectedCalibrationError: { type: number }
        predictionStdDev: { type: number }
        brierSkillScore: { type: ["number", "null"] }
        calibrationSlope: { type: number }
        calibrationIntercept: { type: number }
        promotedAt: { type: string, format: date-time }
      required: [versionId, market, status, displacedVersionId, displacedStatus, walkForwardRunId, foldCount, expectedCalibrationError, predictionStdDev, brierSkillScore, calibrationSlope, calibrationIntercept, promotedAt]
    ModelDemotionResult:
      type: object
      additionalProperties: false
      properties:
        versionId: { type: string }
        market: { type: string, enum: [TB, XBH, WALK, HR] }
        status: { type: string, enum: [RETIRED] }
        demotedAt: { type: string, format: date-time }
      required: [versionId, market, status, demotedAt]
    WalkForwardCalibrationPoint:
      type: object
      additionalProperties: false
      properties:
        bucket: { type: integer, minimum: 0 }
        count: { type: integer, minimum: 0 }
        predictedProbability: { type: ["number", "null"] }
        observedRate: { type: ["number", "null"] }
      required: [bucket, count, predictedProbability, observedRate]
    WalkForwardValidationResult:
      type: object
      additionalProperties: false
      properties:
        walkForwardRunId: { type: string }
        versionId: { type: string }
        market: { type: string, enum: [TB, XBH, WALK, HR] }
        foldCount: { type: integer, minimum: 0 }
        foldResults: { type: array, items: { type: object } }
        overallMetric: { type: ["number", "null"] }
        benchmarkMetric: { type: ["number", "null"] }
        benchmarkBeat: { type: boolean }
        benchmarkMethod: { type: string }
        calibrationMethod: { type: string }
        benchmarkMargin:
          type: ["number", "null"]
          description: |
            Brier margin the model had to beat, derived from the held-out row
            count rather than being a fixed constant.
        calibrationCurve:
          type: array
          description: Ten-bin reliability curve over the pooled out-of-fold probabilities.
          items: { $ref: "#/components/schemas/WalkForwardCalibrationPoint" }
        expectedCalibrationError:
          type: ["number", "null"]
          description: |
            Expected calibration error over the reliability bins. This is the
            quantity the acceptance gate reads.
        expectedCalibrationErrorThreshold: { type: number }
        meanAbsolutePredictionError:
          type: ["number", "null"]
          description: |
            Mean absolute distance between each prediction and its binary label.
            Reported for continuity with the field previously and wrongly named
            calibrationError. Nothing gates on it.
        brierSkillScore: { type: ["number", "null"] }
        predictionStdDev:
          type: ["number", "null"]
          description: Standard deviation of the pooled predicted probabilities, the sharpness guard.
        predictionStdDevThreshold: { type: number }
        failureReasons:
          type: array
          description: Named reasons the run did not pass. Empty on a PASS.
          items: { type: string }
        calibrationPassed: { type: boolean }
        calibrationSlope: { type: ["number", "null"] }
        calibrationIntercept: { type: ["number", "null"] }
        status: { type: string, enum: [PASS, FAIL, INCOMPLETE] }
      required: [walkForwardRunId, versionId, market, foldCount, foldResults, overallMetric, benchmarkMetric, benchmarkBeat, benchmarkMethod, benchmarkMargin, calibrationMethod, calibrationCurve, expectedCalibrationError, expectedCalibrationErrorThreshold, meanAbsolutePredictionError, brierSkillScore, predictionStdDev, predictionStdDevThreshold, failureReasons, calibrationPassed, calibrationSlope, calibrationIntercept, status]
    WalkForwardValidationList:
      type: object
      additionalProperties: false
      properties:
        runs:
          type: array
          items:
            allOf:
              - $ref: "#/components/schemas/WalkForwardValidationResult"
              - type: object
                properties:
                  startedAt: { type: string, format: date-time }
                  finishedAt: { type: ["string", "null"], format: date-time }
                  errorMessage: { type: ["string", "null"] }
                required: [startedAt, finishedAt, errorMessage]
        total: { type: integer, minimum: 0 }
      required: [runs, total]
...