# lib/api-zod/src/generated/types/modelPromotionResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ModelPromotionResultMarket } from './modelPromotionResultMarket';
import type { ModelPromotionResultStatus } from './modelPromotionResultStatus';

export interface ModelPromotionResult {
  versionId: string;
  market: ModelPromotionResultMarket;
  status: ModelPromotionResultStatus;
  /** @nullable */
  displacedVersionId: string | null;
  /** @nullable */
  displacedStatus: string | null;
  walkForwardRunId: string;
  /** @minimum 0 */
  foldCount: number;
  expectedCalibrationError: number;
  predictionStdDev: number;
  /** @nullable */
  brierSkillScore: number | null;
  calibrationSlope: number;
  calibrationIntercept: number;
  promotedAt: Date;
}